+++ packages/react-adapter/src/hooks/useRemoteModule.ts
@@ -1,4 +1,4 @@
-import { loadRemoteMultiVersion } from 'fan-mf-runtime'
+import { loadRemoteMultiVersion } from '@fan-scripts/fan-mf-runtime'
 import React, { useEffect, useState } from 'react'
 
 interface ModuleState {