+++ packages/vue-adapter/src/components/VueRemoteModuleProvider.ts
@@ -10,7 +10,7 @@ import {
   type Component,
   type PropType,
 } from "vue";
-import { loadRemoteMultiVersion } from "fan-mf-runtime";
+import { loadRemoteMultiVersion } from "@fan-scripts/fan-mf-runtime";
 import type {
   MFInstance,
   ReactDOMRoot,