+++ apps/host-vue3-remote/CHANGELOG.md
@@ -4,10 +4,10 @@
 
 ### Patch Changes
 
-- @yunfan/vue-adapter@1.0.1
+- @fan-scripts/vue-adapter@1.0.1
 
 ## 0.0.1
 
 ### Patch Changes
 
-- @yunfan/vue-adapter@1.0.0
+- @fan-scripts/vue-adapter@1.0.0